+++ packages/components-react/src/examples/index.ts
@@ -1,6 +1,7 @@
 export * from './AccordionExample';
 export * from './InlineNotificationExampleActionButton';
 export * from './InlineNotificationExampleEvents';
+export * from './SegmentedControlExample';
 export * from './TableExampleAdvanced';
 export * from './TableExampleBasic';
 export * from './TableExampleSorting';