+++ packages/components/src/components.d.ts
@@ -27,6 +27,7 @@ import { ModalAriaAttributes } from "./components/content/modal/modal-utils";
 import { NumberOfPageLinks, PageChangeEvent } from "./components/navigation/pagination/pagination-utils";
 import { PopoverDirection } from "./components/feedback/popover/popover-utils";
 import { GradientColorTheme, PrevNextButtonJssStyle, ScrollToPosition } from "./components/common/scroller/scroller-utils";
+import { SegmentedControlBackgroundColor, SegmentedControlChangeEvent } from "./components/segmented-control/segmented-control/segmented-control-utils";
 import { DropdownDirection } from "./components/form/select-wrapper/select-wrapper/select-wrapper-utils";
 import { SpinnerAriaAttributes, SpinnerSize } from "./components/feedback/spinner/spinner-utils";
 import { SwitchChangeEvent } from "./components/action/switch/switch";
@@ -653,7 +654,7 @@ export namespace Components {
         /**
           * Preferred direction in which popover should open, given there is enough space in viewport. Otherwise it will be opened in the direction with most available space.
          */
-        "direction": PopoverDirection;
+        "direction"?: PopoverDirection;
     }
     interface PRadioButtonWrapper {
         /**
@@ -687,6 +688,42 @@ export namespace Components {
           * Adapts the color when used on dark background.
          */
         "theme"?: ThemeExtendedElectric;
+    }
+    interface PSegmentedControl {
+        /**
+          * Background color variations
+         */
+        "backgroundColor"?: SegmentedControlBackgroundColor;
+        /**
+          * Adapts the segmented-control color depending on the theme.
+         */
+        "theme"?: Theme;
+        /**
+          * Sets the initial value of the segmented-control.
+         */
+        "value"?: string | number;
+    }
+    interface PSegmentedControlItem {
+        /**
+          * Disables the button. No events will be triggered while disabled state is active.
+         */
+        "disabled"?: boolean;
+        /**
+          * The icon shown.
+         */
+        "icon"?: IconName;
+        /**
+          * A URL path to a custom icon.
+         */
+        "iconSource"?: string;
+        /**
+          * The label text.
+         */
+        "label"?: string;
+        /**
+          * The value of this item which is emitted by the parent element if it becomes selected. This property is **required**.
+         */
+        "value": string | number;
     }
     interface PSelectWrapper {
         /**
@@ -1184,6 +1221,18 @@ declare global {
         prototype: HTMLPScrollerElement;
         new (): HTMLPScrollerElement;
     };
+    interface HTMLPSegmentedControlElement extends Components.PSegmentedControl, HTMLStencilElement {
+    }
+    var HTMLPSegmentedControlElement: {
+        prototype: HTMLPSegmentedControlElement;
+        new (): HTMLPSegmentedControlElement;
+    };
+    interface HTMLPSegmentedControlItemElement extends Components.PSegmentedControlItem, HTMLStencilElement {
+    }
+    var HTMLPSegmentedControlItemElement: {
+        prototype: HTMLPSegmentedControlItemElement;
+        new (): HTMLPSegmentedControlItemElement;
+    };
     interface HTMLPSelectWrapperElement extends Components.PSelectWrapper, HTMLStencilElement {
     }
     var HTMLPSelectWrapperElement: {
@@ -1348,6 +1397,8 @@ declare global {
         "p-popover": HTMLPPopoverElement;
         "p-radio-button-wrapper": HTMLPRadioButtonWrapperElement;
         "p-scroller": HTMLPScrollerElement;
+        "p-segmented-control": HTMLPSegmentedControlElement;
+        "p-segmented-control-item": HTMLPSegmentedControlItemElement;
         "p-select-wrapper": HTMLPSelectWrapperElement;
         "p-select-wrapper-dropdown": HTMLPSelectWrapperDropdownElement;
         "p-spinner": HTMLPSpinnerElement;
@@ -2047,6 +2098,46 @@ declare namespace LocalJSX {
          */
         "theme"?: ThemeExtendedElectric;
     }
+    interface PSegmentedControl {
+        /**
+          * Background color variations
+         */
+        "backgroundColor"?: SegmentedControlBackgroundColor;
+        /**
+          * Emitted when selected element changes.
+         */
+        "onSegmentedControlChange"?: (event: CustomEvent<SegmentedControlChangeEvent>) => void;
+        /**
+          * Adapts the segmented-control color depending on the theme.
+         */
+        "theme"?: Theme;
+        /**
+          * Sets the initial value of the segmented-control.
+         */
+        "value"?: string | number;
+    }
+    interface PSegmentedControlItem {
+        /**
+          * Disables the button. No events will be triggered while disabled state is active.
+         */
+        "disabled"?: boolean;
+        /**
+          * The icon shown.
+         */
+        "icon"?: IconName;
+        /**
+          * A URL path to a custom icon.
+         */
+        "iconSource"?: string;
+        /**
+          * The label text.
+         */
+        "label"?: string;
+        /**
+          * The value of this item which is emitted by the parent element if it becomes selected. This property is **required**.
+         */
+        "value"?: string | number;
+    }
     interface PSelectWrapper {
         /**
           * The description text.
@@ -2436,6 +2527,8 @@ declare namespace LocalJSX {
         "p-popover": PPopover;
         "p-radio-button-wrapper": PRadioButtonWrapper;
         "p-scroller": PScroller;
+        "p-segmented-control": PSegmentedControl;
+        "p-segmented-control-item": PSegmentedControlItem;
         "p-select-wrapper": PSelectWrapper;
         "p-select-wrapper-dropdown": PSelectWrapperDropdown;
         "p-spinner": PSpinner;
@@ -2490,6 +2583,8 @@ declare module "@stencil/core" {
             "p-popover": LocalJSX.PPopover & JSXBase.HTMLAttributes<HTMLPPopoverElement>;
             "p-radio-button-wrapper": LocalJSX.PRadioButtonWrapper & JSXBase.HTMLAttributes<HTMLPRadioButtonWrapperElement>;
             "p-scroller": LocalJSX.PScroller & JSXBase.HTMLAttributes<HTMLPScrollerElement>;
+            "p-segmented-control": LocalJSX.PSegmentedControl & JSXBase.HTMLAttributes<HTMLPSegmentedControlElement>;
+            "p-segmented-control-item": LocalJSX.PSegmentedControlItem & JSXBase.HTMLAttributes<HTMLPSegmentedControlItemElement>;
             "p-select-wrapper": LocalJSX.PSelectWrapper & JSXBase.HTMLAttributes<HTMLPSelectWrapperElement>;
             "p-select-wrapper-dropdown": LocalJSX.PSelectWrapperDropdown & JSXBase.HTMLAttributes<HTMLPSelectWrapperDropdownElement>;
             "p-spinner": LocalJSX.PSpinner & JSXBase.HTMLAttributes<HTMLPSpinnerElement>;