+++ packages/components/src/components/feedback/inline-notification/inline-notification.spec.ts
@@ -1,10 +1,10 @@
-import * as validationUtils from '../../../utils/validation';
+import * as throwIfValueIsInvalidUtils from '../../../utils/validation/throwIfValueIsInvalid';
 import { InlineNotification } from './inline-notification';
 import { INLINE_NOTIFICATION_STATES } from './inline-notification-utils';
 
 describe('componentWillRender', () => {
   it('should call throwIfValueIsInvalid() with correct parameters', () => {
-    const spy = jest.spyOn(validationUtils, 'throwIfValueIsInvalid');
+    const spy = jest.spyOn(throwIfValueIsInvalidUtils, 'throwIfValueIsInvalid');
     const component = new InlineNotification();
     component.host = document.createElement('p-inline-notification');
     component.host.attachShadow({ mode: 'open' });